Moves remoteFollow(uri:) request to Accounts
